Clean up save_model and load_weights
I use numpy.savez to save the weights of the model. When save_model is called, the hyperparameterswill be saved as 'filename' and the weights as the var weights_file. If weights_file is specified, then use numpy.savez to get_all_param_values of network. 

When load_model_from_json is called, model_fp loads the hyperparameters. If weights_fp is specified, then load load_weights which will use set_all_param_values based on numpy.savz output.